requirements.txt
openai
gspread
python-dotenv
playwright
